TagsInput is now smarter with input
Issue: Entering strings separated by ‘,’ was broken. Users could enter ‘,’ and it would result in two empty tags, or enter ‘a’, after which ‘a,a’ which would result in 3 ‘a’ in the input.

New logic fixes that and also filters and adds only the new unique and new tags. So if you have ‘a,b,c’ and you enter ‘d,a,b,c’ only ‘d’ would be added etc.